requirements.txt
python-dotenv
django
djangorestframework
llama_index
llama-index-vector-stores-postgres
llama-index-llms-anthropic
llama-index-llms-google-genai 
llama-index-embeddings-google-genai
sqlalchemy
opik
litellm
google-generativeai
pandas
beautifulsoup4
requests
jsonschema


